+++ tests/TestCase/Utility/BBCodeTest.php
@@ -55,6 +55,7 @@ class BBCodeTest extends TestCase
     {
         $expected = '<p>Some para text</p>' . PHP_EOL .
             '<!-- read-more -->' . PHP_EOL .
+            '<hr />' . PHP_EOL .
             '<span>Some span text</span>' . PHP_EOL .
             '<div class="ratio ratio-16x9"><iframe allowfullscreen="allowfullscreen" height="480" src="https://www.youtube.com/embed/bL_CJKq9rIw" width="640"></iframe></div>' . PHP_EOL .
             '<div>Some div text</div>' . PHP_EOL;
@@ -62,6 +63,7 @@ class BBCodeTest extends TestCase
         ob_start();
         echo '<p>Some para text</p>' . PHP_EOL;
         echo '[readmore /]' . PHP_EOL;
+        echo '[hr /]' . PHP_EOL;
         echo '<span>Some span text</span>' . PHP_EOL;
         echo '[youtube]bL_CJKq9rIw[/youtube]' . PHP_EOL;
         echo '<div>Some div text</div>' . PHP_EOL;
@@ -89,6 +91,18 @@ class BBCodeTest extends TestCase
         echo '<div>Some div text</div>' . PHP_EOL;
         $result = $this->BBCode->remove(ob_get_clean() ?: '');
         $this->assertSame($expected, $result);
+    }
+
+    /**
+     * Tests for `hr()` method
+     * @uses \MeTools\Utility\BBCode::hr()
+     * @test
+     */
+    public function testHr(): void
+    {
+        $this->assertSame('<hr />', $this->BBCode->hr('[hr]'));
+        $this->assertSame('<hr />', $this->BBCode->hr('[hr/]'));
+        $this->assertSame('<hr />', $this->BBCode->hr('[hr /]'));
     }
 
     /**
@@ -149,5 +163,16 @@ class BBCodeTest extends TestCase
         ] as $text) {
             $this->assertSame($expected, $this->BBCode->youtube($text));
         }
+
+        $expected = $this->Html->youtube('-YcwR89cfao?t=62');
+        foreach ([
+            '[youtube]https://youtu.be/-YcwR89cfao?t=62[/youtube]',
+            '[youtube]-YcwR89cfao?t=62[/youtube]',
+        ] as $text) {
+            $this->assertSame($expected, $this->BBCode->youtube($text));
+        }
+
+        $expected = $this->Html->youtube('bL_CJKq9rIw');
+        $this->assertSame($expected, $this->BBCode->youtube('<p>[youtube]bL_CJKq9rIw[/youtube]</p>'));
     }
 }