+++ tests/TestCase/View/Helper/FormHelperTest.php
@@ -311,7 +311,7 @@ class FormHelperTest extends HelperTestCase
     {
         $this->assertFalse($this->Helper->isInline());
 
-        $expected = '<form method="post" accept-charset="utf-8" class="align-items-center g-3 my-class row row-cols-lg-auto" action="/">';
+        $expected = '<form method="post" accept-charset="utf-8" class="my-class" action="/">';
         $result = $this->Helper->createInline(null, ['class' => 'my-class']);
         $this->assertSame($expected, $result);